+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- **`typing-extensions` is not being used; it is not a dependency**
+
 ## [0.6.0-beta.1] - 2022-07-10
 
 ### Added